tambah riwayat aktivitas

+++ database/migrations/2021_10_07_055047_create_bahans_table.php
@@ -19,8 +19,8 @@ class CreateBahansTable extends Migration
             $table->string('nama_bahan');
             $table->integer('jumlah_bahan');
             $table->string('satuan_bahan');
-            $table->bigInteger('harga_bahan');
-            $table->bigInteger('harga_satuan')->nullable();
+            $table->double('harga_bahan');
+            $table->double('harga_satuan')->nullable();
             $table->timestamps();
         });
     }
